+++ R/health_data_scotland.R
@@ -4,9 +4,10 @@
 health_data_scotland <- function(...) {
   requireNamespace("sf", quietly = TRUE)
 
-  log_info("Creating data objects")
+  log_info("Creating initial objects")
   data <- create_data_objects(load_processed_data())
-  log_info("Created data objects")
+  map <- create_map_object(data)
+  log_info("Created initial objects")
 
   value_boxes <- data |>
     purrr::imap(function(x, nm) {
@@ -32,16 +33,7 @@ health_data_scotland <- function(...) {
       title = "Map",
       full_screen = TRUE,
       height = "600px",
-      nav_panel(
-        title = "Interactive map",
-        map_UI(
-          id = "map",
-          boards = get_sf("board") |>
-            as_tibble() |>
-            select("HBName", "ID") |>
-            deframe()
-        )
-      )
+      map[["ui"]]()
     ),
     nav_menu(
       title = "Summary",
@@ -83,7 +75,7 @@ health_data_scotland <- function(...) {
       glue("Launching HealthDataScotland v{packageVersion('HealthDataScotland')}")
     )
     log_info(glue::glue_collapse(list.files(), "\n"))
-    map_server("map", data, get_sf("board"))
+    map[["server"]](data)
     data[["gp"]][["server"]]()
     data[["hospital"]][["server"]]()
   }